requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.4.0
requests==2.31.0
openai==0.27.8
moviepy==1.0.3
ffmpeg-python==0.2.0
Werkzeug==2.3.7
bcrypt==4.0.1
itsdangerous==2.1.2
beautifulsoup4==4.12.2

